# Fix handling of extended characters in .stignore on macOS