--- a/lib/ignore/ignore.go
+++ b/lib/ignore/ignore.go
@@ -248,7 +248,7 @@ func (m *Matcher) Match(file string) (result ignoreresult.R) {
 	// allow skipping matched directories or not. As soon as we hit an
 	// exclude pattern (with some exceptions), we can't skip directories
 	// anymore.
-	file = filepath.ToSlash(file)
+	file = osutil.NormalizedFilename(file)
 	var lowercaseFile string
 	canSkipDir := true
 	for _, pattern := range m.patterns {