### Purpose

Normalize pathnames before comparing to ignore patterns. (issue #9597).

Added to the existing unit test. The existing unit tests covers extended character patterns, but it is when a NFD filename is matched against a NFC pattern that the an expected match did not occur.

Made the unit test more robust by passing in native-format pathnames every time Matcher.Match is invoked, to uncover potential future similar bugs.

### Testing

Updated unit test fails without fix, passes with fix. (On macOS at least).

Benchmark results:

```
Before:
BenchmarkMatch-8                 2180224               538.7 ns/op
BenchmarkMatchCached-8           9932952               119.6 ns/op

BenchmarkMatch-8                 2192263               534.9 ns/op
BenchmarkMatchCached-8           9533098               119.8 ns/op

BenchmarkMatch-8                 2171739               536.4 ns/op
BenchmarkMatchCached-8           9956037               120.5 ns/op

BenchmarkMatch-8                 2192551               538.2 ns/op
BenchmarkMatchCached-8           9385444               121.7 ns/op

BenchmarkMatch-8                 2140863               539.8 ns/op
BenchmarkMatchCached-8           9518190               122.1 ns/op

BenchmarkMatch-8                 2189522               531.4 ns/op
BenchmarkMatchCached-8           9751231               118.8 ns/op

BenchmarkMatch-8                 2192011               536.6 ns/op
BenchmarkMatchCached-8           9524190               119.8 ns/op


After:
BenchmarkMatch-8                 2164195               540.4 ns/op
BenchmarkMatchCached-8           9753216               119.9 ns/op

BenchmarkMatch-8                 2177550               540.8 ns/op
BenchmarkMatchCached-8           9671676               120.2 ns/op

BenchmarkMatch-8                 2171844               541.7 ns/op
BenchmarkMatchCached-8           9651882               119.6 ns/op

BenchmarkMatch-8                 2164958               540.8 ns/op
BenchmarkMatchCached-8           9804344               119.4 ns/op

BenchmarkMatch-8                 2141856               542.0 ns/op
BenchmarkMatchCached-8           9493785               120.3 ns/op

BenchmarkMatch-8                 2168912               539.3 ns/op
BenchmarkMatchCached-8           9555124               120.2 ns/op

BenchmarkMatch-8                 2171197               540.3 ns/op
BenchmarkMatchCached-8           9690603               119.6 ns/op
```

This shows a 0.87% mean performance degradation on `BenchmarkMatch`

I have considered how to minimise this. The normalisation could be applied only if the any of the ignore patterns contain any extended characters (detected by whether NFC-form matches NFD-form). Interested in others' thoughts on whether this needs to be optimised?

### Screenshots

N/A

### Documentation

N/A

## Authorship

Your name and email will be added automatically to the AUTHORS file
based on the commit metadata.