@calmh 2024-06-29
Looks reasonable to me. The bulk of the changes to the test code are unnecessary though, calling NativeFilename on hard coded ASCII strings.