@SimonPickup 2024-06-29
> Looks reasonable to me. The bulk of the changes to the test code are unnecessary though, calling NativeFilename on hard coded ASCII strings.

Yeah I went back and forth about whether to update them all. Yes they're empriically unnecessary, though technically correct, and help protect against hypothetical future bugs where native/normalised form could have other variations. And there's a cost to test readability.

I deliberately put this in a separate commit specifically so it can be reverted if you prefer. I don't object either way.