@imsodin 2024-07-02
My gut feeling is that this is the wrong way around, and we should denormalize/nativefy the ignore patterns instead, however I might be confusing things so let me try to explain (or showcase the confusion):

BEP is normalized, however we convert everything to native format - so any strings coming from the DB (file infos, ...) should be NFD on macos. Macos filesystem uses NFD so anything from fs interactions is NFD. So my initial confusion is: Where does NFC come into this? Some reading (very inconclusively) suggest not everything is NFD, just filesystem. So I presume text entered through the browser/gui/api ends up as NFC in our ignore file. Same for text editors possibly (tested that bash shenanigans on macos produce an NFC "ä").  
Anyway, given we use NFD for all file info in DB, it seems like it would make sense to use NFD for (parsed) ignore patterns as well. Probably call `osutil.NativeFilename` early in `parseLine`.

Does that make any sense or can you point out where I confuse things? :)