@imsodin 2024-07-03
Oh I by no means am sure that anything coming from FS is NFD normalised. I just know that syncthing internal data/strings/filenames are nfd. so a minimal fix for the issue you describe seems to be to ensure ignores are also nfd. if it then turns out that some fs aren't normalised, that seems like an additional, somewhat independent issue.