@calmh 2024-07-03
Iirc we don't actually normalise for internal work, we just normalise on the wire and in the database. I think the solution here is proper, as far as I understand how it's supposed to work.

(Maybe we should always use nfc or whatever internally and just convert at the outermost layer, though.)