@bt90 2024-07-03
> Maybe we should always use nfc or whatever internally and just convert at the outermost layer, though.

Yeah. This whole normalization thing is rather expensive.