@imsodin 2024-07-05
> (Maybe we should always use nfc or whatever internally and just convert at the outermost layer, though.)

Afaik that's what we do already (modulo the stuff I missed):

1. Wire/protocol  
Everything going to the wire is NFC normalized. On macos it gets converted to NFD from the wire, on the rest left alone (so still NFC).
2. Filesystem  
On the filesystem layer itself afaik we don't do anything about normalization. However while scanning/putting things into DB we do normalize - NFD on macos, NFC everywhere else.

I don't see any other points of entry for filenames, are there?

So from the above I'd think we are able to trust that input to ignores is normalized. Looking at the scanning code a bit closer though shows that's not the case: We test "ignoredness" before normalization. However it seems like fixing/changing that, and then just deal with normalized filenames only internally seems cleaner than re-normalize whereever we need to be sure things are normalized (with "whereever" == ignores right now, slight hyperbole there :P ).